+++ src/components/Main/Metrics/MetricsEmployee/MetricsEmployee.jsx
@@ -1,56 +1,65 @@
-import { useEffect, useMemo, useState } from "react";
+import { useContext, useEffect, useMemo, useState } from "react";
 import Carbon from "./Carbon/Carbon.jsx";
 import Employee from "./Employee/Employee.jsx";
 import Office from "./Office/Office.jsx";
 import BarChartComponent from "./BarChart/BarChart.jsx";
 import Workstation from "./Workstation/Workstation.jsx";
 import PieChartComponent from "./PieChart/PieChart.jsx";
+import { apiCallForMetrics, MetricsContext } from "../../../../store/metrics-context.jsx";
 import { $MetricsEmployee } from "./MetricsEmployee.styles.jsx";
-import {
-  apiCallToGetEmployeeMetric,
-  apiCallToGetEmployeeCarbonFootprint,
-} from "../../../../api/Api.jsx";
+import { apiCallToGetEmployeeCarbonFootprint } from "../../../../api/Api.jsx";
 
 export default function MetricsEmployee() {
+  const { currentStep } = useContext(MetricsContext);
   const [employeeMetric, setEmployeeMetric] = useState({});
 
   useEffect(() => {
-    const metrics = apiCallToGetEmployeeMetric(0);
-    setEmployeeMetric((_prevMetrics) => metrics);
+    async function fetchData() {
+      try {
+        await apiCallForMetrics(currentStep).then(resData => {
+          const { id, corporate_key, email, name, surname, role, carbonLimit, location, workstation } = resData;
+          setEmployeeMetric((_prevMetrics) => ({ id, corporate_key, email, name, surname, role, carbonLimit, location, workstation }));
+        });
+      } catch (error) {
+        console.log(error);
+        //TODO
+      }
+    }
+
+    fetchData()
   }, []);
 
   const carbonFootprint = useMemo(() => {
     return apiCallToGetEmployeeCarbonFootprint(0);
   }, [employeeMetric]);
 
-  const {
-    id,
-    corporate_key,
-    email,
-    name,
-    surname,
-    role,
-    carbon_limit,
-    location,
-    office_id,
-    workstation_id,
-  } = employeeMetric;
-  const employee = {
-    corporate_key,
-    email,
-    name,
-    surname,
-    role,
-    location,
-  };
+  function getEmployeeData() {
+    return {
+      corporate_key: employeeMetric.corporate_key,
+      email: employeeMetric.email, 
+      name: employeeMetric.name, 
+      surname: employeeMetric.surname,
+      role: employeeMetric.role && employeeMetric.role.name,
+      city: employeeMetric.location && employeeMetric.location.city,
+      country: employeeMetric.location && employeeMetric.location.country && employeeMetric.location.country.name
+    }
+  }
+
+  function getOfficeData() {
+    return {}
+  }
+
+  function getWorkstationData() {
+    return {}
+  }
 
   return (
     <$MetricsEmployee>
-      <Employee employee={employee} />
-      <Office office={office_id} />
-      <Workstation workstation={workstation_id} />
-      <Carbon employeeId={id} carbonFootprint={carbonFootprint} carbonLimit={carbon_limit} />
-      <BarChartComponent carbonFootprint={carbonFootprint} carbonLimit={carbon_limit} />
+      <Employee employee={getEmployeeData()} />
+      <Office office={employeeMetric.office_id} />
+      <Workstation workstation={employeeMetric.workstation} />
+      <Carbon employeeId={employeeMetric.id} carbonFootprint={carbonFootprint} carbonLimit={employeeMetric.carbonLimit} />
+      <BarChartComponent carbonFootprint={carbonFootprint} carbonLimit={employeeMetric.carbonLimit} />
       <PieChartComponent carbonFootprint={carbonFootprint} />
     </$MetricsEmployee>
   );